fixes warning

+++ pkg/ant.ecs/src/ecs_component.cpp
@@ -4,7 +4,7 @@
 extern "C" int
 luaopen_ecs_components(lua_State *L) {
     const auto& components = ecs::decl::components;
-    lua_createtable(L, components.size(), components.size());
+    lua_createtable(L, (int)components.size(), (int)components.size());
     for (size_t i = 0; i < components.size(); ++i) {
         lua_pushlstring(L, components[i].name.data(), components[i].name.size());
         if (components[i].size != 0) {
